+++ packages/editor/src/components/post-taxonomies/hierarchical-term-selector.js
@@ -13,11 +13,15 @@ import {
 	Flex,
 	FlexItem,
 	SearchControl,
+	Spinner,
 	privateApis as componentsPrivateApis,
 } from '@wordpress/components';
 import { useDispatch, useSelect } from '@wordpress/data';
 import { useDebounce } from '@wordpress/compose';
-import { store as coreStore } from '@wordpress/core-data';
+import {
+	store as coreStore,
+	privateApis as coreDataPrivateApis,
+} from '@wordpress/core-data';
 import { speak } from '@wordpress/a11y';
 import { decodeEntities } from '@wordpress/html-entities';
 
@@ -28,6 +32,9 @@ import { buildTermsTree } from '../../utils/terms';
 import { store as editorStore } from '../../store';
 import { unlock } from '../../lock-unlock';
 
+const { normalizeTextString } = unlock( componentsPrivateApis );
+const { RECEIVE_INTERMEDIATE_RESULTS } = unlock( coreDataPrivateApis );
+
 /**
  * Module Constants
  */
@@ -37,11 +44,10 @@ const DEFAULT_QUERY = {
 	order: 'asc',
 	_fields: 'id,name,parent',
 	context: 'view',
+	[ RECEIVE_INTERMEDIATE_RESULTS ]: true,
 };
 const MIN_TERMS_COUNT_FOR_FILTER = 8;
 const EMPTY_ARRAY = [];
-
-const { normalizeTextString } = unlock( componentsPrivateApis );
 
 /**
  * Sort Terms by Selected.
@@ -186,14 +192,14 @@ export function HierarchicalTermSelector( { slug } ) {
 
 			return {
 				hasCreateAction: _taxonomy
-					? post._links?.[
+					? !! post._links?.[
 							'wp:action-create-' + _taxonomy.rest_base
-					  ] ?? false
+					  ]
 					: false,
 				hasAssignAction: _taxonomy
-					? post._links?.[
+					? !! post._links?.[
 							'wp:action-assign-' + _taxonomy.rest_base
-					  ] ?? false
+					  ]
 					: false,
 				terms: _taxonomy
 					? getEditedPostAttribute( _taxonomy.rest_base )
@@ -410,7 +416,7 @@ export function HierarchicalTermSelector( { slug } ) {
 
 	return (
 		<Flex direction="column" gap="4">
-			{ showFilter && (
+			{ showFilter && ! loading && (
 				<SearchControl
 					__next40pxDefaultSize
 					__nextHasNoMarginBottom
@@ -419,6 +425,17 @@ export function HierarchicalTermSelector( { slug } ) {
 					value={ filterValue }
 					onChange={ setFilter }
 				/>
+			) }
+			{ loading && (
+				<Flex
+					justify="center"
+					style={ {
+						// Match SearchControl height to prevent layout shift.
+						height: '40px',
+					} }
+				>
+					<Spinner />
+				</Flex>
 			) }
 			<div
 				className="editor-post-taxonomies__hierarchical-terms-list"