+++ packages/icons/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+## 10.30.0 (2025-09-03)
+
+## 10.29.0 (2025-08-20)
+
+-  Breaking change: deprecate `moreHorizontalMobile` icon. This was added in error, and `moreHorizontal` should be used instead. ([#71172](https://github.com/WordPress/gutenberg/pull/71172))
+
 ## 10.28.0 (2025-08-07)
 
 -   Add new `dashboard` icon ([#70300](https://github.com/WordPress/gutenberg/pull/70300)).